+++ Imaj.Data/Repositories/Repository.cs
@@ -36,7 +36,7 @@ namespace Imaj.Data.Repositories
             return await _dbSet.ToListAsync();
         }
 
-        public async Task<T?> GetByIdAsync(int id)
+        public async Task<T?> GetByIdAsync(decimal id)
         {
             return await _dbSet.FindAsync(id);
         }
